304-Web/Standard-Login-and-Registration/requirements.txt
Flask==3.1.0
Flask_Limiter==3.8.0
matplotlib==3.9.2
pandas==2.2.3
prophet==1.1.6
pyodbc==5.2.0
python-dotenv==1.0.1
Requests==2.32.3
Werkzeug==3.1.3
